Django/requirements.txt
aioitertools==0.12.0
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
autopep8==2.3.1
bcrypt==4.2.0
blinker==1.8.2
brewer2mpl==1.4.1
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
config==0.5.1
cycler==0.12.1
Django==5.1.4
djangorestframework==3.15.2
dnspython==2.7.0
emoji==2.14.0
fonttools==4.54.1
idna==3.10
itsdangerous==2.2.0
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.7
latexcodec==3.0.0
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numpy==1.26.4
orjson==3.10.10
packaging==24.1
palettable==3.3.3
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
psutil==6.1.0
pybtex==0.24.0
pycodestyle==2.12.1
pycparser==2.22
pydantic_core==2.23.4
pydash==8.0.3
pymongo==4.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
scipy==1.14.1
sentinels==1.0.0
six==1.16.0
spglib==2.5.0
sqlparse==0.5.1
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
uncertainties==3.2.2
urllib3==2.2.3
Werkzeug==3.0.6
